bizosaas-brain-core/brain-gateway/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx>=0.27.0,<0.28.0
pydantic>=2.8.2,<2.9.0
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.1.0
strawberry-graphql[fastapi]==0.219.0
temporalio==1.4.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
hvac==2.1.0
psutil==5.9.8
PyJWT>=2.9.0,<3.0.0
clerk-backend-api==1.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-prometheus==0.43b0
schemathesis==3.31.0
pytest==8.3.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
faker==20.1.0
anyio==4.3.0
redis==5.0.1
razorpay==1.4.1
stripe==8.5.0
